product serializers separated
