+++ ecommerce_project/myapp/serializers/__init__.py
@@ -1,0 +1,1 @@
+from .product_serializers import *
